backend/enhanced_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.25.0
python-multipart==0.0.6
# Amazon Product Advertising API
python-amazon-paapi>=5.0.0
